Bugfix on ending drag on same space as drag started
Fixed bug where a drag & drop that began & ended on the same square would result in that square reporting a null email (a consequence of storing the dragged email  in the destination square and then cleaning out the source square without considering that they could be the same square).

Also WIP on zooming in & out on emails, interrupted partway thru after this bug was found.